refactor: GameController에서 CarList를 다루는 메소드들은 CarListService 클래스 생성시켜 분리
